% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_departments.R
\name{met_departments}
\alias{met_departments}
\title{List The Met departments}
\usage{
met_departments()
}
\value{
A tibble with columns:
\describe{
\item{department_id}{Integer department ID}
\item{display_name}{Department display name}
}
}
\description{
Returns all departments available in The Met Collection API.
}
\examples{
\dontrun{
met_departments()
}
}
