% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_object.R
\name{met_object}
\alias{met_object}
\title{Retrieves one piece of art from the collection}
\usage{
met_object(object_id)
}
\arguments{
\item{object_id}{ID of a single object (piece of art)}
}
\value{
A one-row tibble or an empty tibble if nothing is found
}
\description{
Retrieves one piece of art from the collection
}
