% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_objects.R
\name{met_objects}
\alias{met_objects}
\title{Retrieves multiple pieces of art from the collection}
\usage{
met_objects(object_ids)
}
\arguments{
\item{object_ids}{A vector of object IDs}
}
\value{
Tibble with one row per object. Errors if \code{object_ids} is empty.
}
\description{
Retrieves multiple pieces of art from the collection
}
