% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_tidy_object.R
\name{met_tidy_object}
\alias{met_tidy_object}
\title{Creating a clean tibble from a single object's JSON output}
\usage{
met_tidy_object(x)
}
\arguments{
\item{x}{Raw JSON from the API for a single object}
}
\value{
A one row tibble
}
\description{
Creating a clean tibble from a single object's JSON output
}
